refactor: reestrutura pastas do projeto para organização por camada
- Remove pasta shared/ e distribui em camadas na raiz do src/:
  - shared/components/ → components/
  - shared/hooks/ → hooks/
  - shared/api/ → lib/api.ts
  - shared/validators/ → lib/validators.ts
  - shared/interfaces/ → types/
  - shared/context/ → context/
  - shared/utils/ → utils/
  - i18n.ts → lib/i18n.ts
- Renomeia rota tvShows → tv-shows (kebab-case, padrão de URL)
- Move home.style.scss para co-localizado em components/mediaList/
- Atualiza todos os imports para nova estrutura (@/components, @/hooks, @/lib, @/types, @/utils, @/context)
