+++ src/app/layout.tsx
@@ -3,9 +3,9 @@ import "../styles/globals.css";
 import "primeicons/primeicons.css";
 import { SkeletonTheme } from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
-import { AppProvider } from "@/shared/context/context";
-import { ErrorBoundary } from "@/shared/components/errorBoundary/ErrorBoundary";
-import { Header } from "@/shared/components/header/header";
+import { AppProvider } from "@/context/context";
+import { ErrorBoundary } from "@/components/errorBoundary/ErrorBoundary";
+import { Header } from "@/components/header/header";
 
 export const dynamic = 'force-dynamic';
