+++ src/components/banner/banner.style.scss
