+++ src/components/imageMovie/imageMovie.style.scss
