+++ src/components/layoutComponent/detailsContent.tsx
