+++ src/components/layoutComponent/layout.style.scss
