+++ src/components/modal/modal.styled.scss
