+++ src/components/skeletonLoading/skeletonBanner.tsx
