+++ src/components/skeletonLoading/skeletonDetails.tsx
