+++ src/components/skeletonLoading/skeletonDetailsActors.tsx
