+++ src/components/skeletonLoading/skeletonSlider.tsx
