+++ src/components/sliderActors/sliderActors.style.scss
