+++ src/context/context.tsx
@@ -6,7 +6,7 @@ import {
     useEffect,
     useState
 } from "react";
-import i18n from '../../i18n';
+import i18n from '@/lib/i18n';
 
 type Props = {
     language: string;
