+++ src/types/IActorDetails.ts
