+++ src/types/IImage.ts
