+++ src/utils/formatGenres.ts
